requirements.txt
fastapi
uvicorn
openai
PyPDF2
requests
beautifulsoup4
faiss-cpu
langchain>=0.2.0
langchain-openai
langchain-community
anthropic
google-generativeai
tiktoken
python-dotenv
